Dynamic tenant OAuth configuration
This PR adds the ability to configure OAuth providers per tenant. This is exposed through the tenantOAuths service. When a user attempts to login, they need to provide a `tenantId` in the body of the request. This is used to lookup the OAuth configuration for that tenant. If the tenant is not found, the request is rejected. If the tenant is found, the OAuth configuration is used to redirect the user to the OAuth provider. Once the user has authenticated with the OAuth provider, they are redirected back to the server. If the user is not found, they are created. If the user is found, they are updated. The user is then logged in and a JWT is returned to the client.